Created admin actions

+++ src/main/java/fi/tuni/bloggadeback/BlogPostRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface BlogPostRepository extends CrudRepository<BlogPost, Integer> {
+public interface BlogPostRepository extends CrudRepository<BlogPost, Long> {
     List<BlogPost> findByBlogTitleContainingIgnoreCase(String keyword);
 }
